requirements.txt
monai>=1.5.0
nibabel>=5.2.0
peft>=0.11.0
torch>=2.3.0
tqdm>=4.66.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0.0
